requirements.txt
Flask==1.0.2
gunicorn==19.9.0
pylint==2.3.1
google-cloud-vision==1.0.0
